Fix TCP stream framing issue
Because of the nature of TCP protocol we receive consecutive IO buffer
snapshots - not separate messages. That means that we need to join
incomming buffers and then split it into separate WireFrames.

Closes ONAP-312
Issue-ID: DCAEGEN2-601

+++ hv-collector-client-simulator/src/main/kotlin/org/onap/dcae/collectors/veshv/simulators/xnf/config/ClientConfiguration.kt
@@ -27,4 +27,4 @@ data class ClientConfiguration(
         val vesHost: String,
         val vesPort: Int,
         val security: ClientSecurityConfiguration,
-        val messagesAmount: Int)
+        val messagesAmount: Long)
